chore: migrate testing environment from jsdom to happy-dom and update related configurations
- Updated testing environment in vitest.config.ts to use happy-dom.
- Adjusted references in documentation and code comments to reflect the change from jsdom to happy-dom.
- Replaced jsdom dependency with happy-dom in package.json and pnpm-lock.yaml.
- Enhanced test setup to accommodate happy-dom's limitations, including a polyfill for MessageChannel.
- Removed outdated guidelines file that was no longer relevant.

This change improves compatibility with headless testing environments and optimizes test performance.
