+++ package.json
@@ -137,7 +137,7 @@
     "csstype": "^3.1.2",
     "dedent": "^0.7.0",
     "husky": "^6.0.0",
-    "jsdom": "^29.1.1",
+    "happy-dom": "^20.11.1",
     "lint-staged": "^17.0.8",
     "markdown-table": "^3.0.3",
     "npm-run-all": "^4.1.5",
