+++ src/components/navigation/Tabs/Tabs.test.tsx
@@ -731,398 +731,6 @@ describe('<Tabs />', () => {
     });
   });
 
-  describe('Title Editing', () => {
-    it('should enter edit mode on double-click when isEditable', async () => {
-      const user = userEvent.setup();
-      const { getByRole } = renderWithRoot(
-        <Tabs defaultActiveKey="tab1" onTitleChange={vi.fn()}>
-          <Tab key="tab1" isEditable title="Tab 1">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      const tab = getByRole('tab');
-      const titleSpan = tab.querySelector('span');
-
-      expect(titleSpan).toBeInTheDocument();
-
-      await user.dblClick(titleSpan!);
-
-      await waitFor(() => {
-        expect(getByRole('textbox')).toBeInTheDocument();
-      });
-
-      expect(getByRole('textbox')).toHaveValue('Tab 1');
-    });
-
-    it('should call onTitleChange when Enter is pressed', async () => {
-      const user = userEvent.setup();
-      const handleTitleChange = vi.fn();
-      const { getByRole } = renderWithRoot(
-        <Tabs defaultActiveKey="tab1" onTitleChange={handleTitleChange}>
-          <Tab key="tab1" isEditable title="Tab 1">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      const tab = getByRole('tab');
-      const titleSpan = tab.querySelector('span');
-
-      await user.dblClick(titleSpan!);
-
-      // Wait for edit mode
-      await waitFor(() => {
-        expect(getByRole('textbox')).toBeInTheDocument();
-      });
-
-      const input = getByRole('textbox') as HTMLInputElement;
-
-      // Change value using fireEvent and press Enter
-      await act(async () => {
-        fireEvent.change(input, { target: { value: 'New Title' } });
-        fireEvent.keyDown(input, { key: 'Enter' });
-      });
-
-      await waitFor(() => {
-        expect(handleTitleChange).toHaveBeenCalledWith('tab1', 'New Title');
-      });
-    });
-
-    it('should cancel editing when Escape is pressed', async () => {
-      const user = userEvent.setup();
-      const handleTitleChange = vi.fn();
-      const { getByRole, queryByRole } = renderWithRoot(
-        <Tabs defaultActiveKey="tab1" onTitleChange={handleTitleChange}>
-          <Tab key="tab1" isEditable title="Tab 1">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      const tab = getByRole('tab');
-      const titleSpan = tab.querySelector('span');
-
-      await user.dblClick(titleSpan!);
-
-      // Wait for edit mode
-      await waitFor(() => {
-        expect(getByRole('textbox')).toBeInTheDocument();
-      });
-
-      const input = getByRole('textbox');
-
-      await act(async () => {
-        await user.type(input, 'Changed{Escape}');
-      });
-
-      await waitFor(() => {
-        expect(queryByRole('textbox')).not.toBeInTheDocument();
-      });
-
-      expect(handleTitleChange).not.toHaveBeenCalled();
-    });
-
-    it('should not submit empty title', async () => {
-      const user = userEvent.setup();
-      const handleTitleChange = vi.fn();
-      const { getByRole } = renderWithRoot(
-        <Tabs defaultActiveKey="tab1" onTitleChange={handleTitleChange}>
-          <Tab key="tab1" isEditable title="Tab 1">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      const tab = getByRole('tab');
-      const titleSpan = tab.querySelector('span');
-
-      await user.dblClick(titleSpan!);
-
-      // Wait for edit mode
-      await waitFor(() => {
-        expect(getByRole('textbox')).toBeInTheDocument();
-      });
-
-      const input = getByRole('textbox');
-
-      await act(async () => {
-        await user.clear(input);
-        await user.type(input, '   {Enter}');
-      });
-
-      expect(handleTitleChange).not.toHaveBeenCalled();
-    });
-
-    it('should trim title on submit', async () => {
-      const user = userEvent.setup();
-      const handleTitleChange = vi.fn();
-      const { getByRole } = renderWithRoot(
-        <Tabs defaultActiveKey="tab1" onTitleChange={handleTitleChange}>
-          <Tab key="tab1" isEditable title="Tab 1">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      const tab = getByRole('tab');
-      const titleSpan = tab.querySelector('span');
-
-      await user.dblClick(titleSpan!);
-
-      // Wait for edit mode
-      await waitFor(() => {
-        expect(getByRole('textbox')).toBeInTheDocument();
-      });
-
-      const input = getByRole('textbox') as HTMLInputElement;
-
-      // Change value with whitespace and press Enter - should trim
-      await act(async () => {
-        fireEvent.change(input, { target: { value: '  Trimmed Title  ' } });
-        fireEvent.keyDown(input, { key: 'Enter' });
-      });
-
-      await waitFor(() => {
-        expect(handleTitleChange).toHaveBeenCalledWith('tab1', 'Trimmed Title');
-      });
-    });
-
-    it('forwards an explicit tab tooltip to InlineInput and suppresses Item tooltip for editable tabs', () => {
-      const { getByTestId, getByRole } = renderWithRoot(
-        <Tabs defaultActiveKey="tab1" onTitleChange={vi.fn()}>
-          <Tab key="tab1" isEditable title="Tab 1" tooltip="Click to rename">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      // The tab button no longer participates in tooltip handling — the
-      // InlineInput root owns it. Both the InlineInput root and the tab
-      // button should remain in the DOM with no double-wrapping side effects.
-      const tab = getByRole('tab');
-      const inlineInput = getByTestId('InlineInput');
-
-      expect(tab).toContainElement(inlineInput);
-      // No regression in the editing flow: dblclick still enters edit mode.
-      expect(inlineInput).toHaveAttribute('data-qa', 'InlineInput');
-    });
-
-    // Regression: a "Rename" item inside the tab menu used to mount the
-    // inline-edit input only to have it stolen-focus + committed-out by the
-    // closing Menu popover's `<FocusScope restoreFocus>`. The user clicked
-    // Rename and nothing visibly happened. InlineInput's grace period plus
-    // TabButton's refocus pass keep the input mounted and focused across the
-    // overlay's exit.
-    it('keeps the rename input mounted and focused when triggered from the tab menu', async () => {
-      const user = userEvent.setup();
-      const handleTitleChange = vi.fn();
-      const { findByRole, getByRole, queryByRole } = renderWithRoot(
-        <Tabs
-          isEditable
-          defaultActiveKey="tab1"
-          menu={<Menu.Item key="rename">Rename</Menu.Item>}
-          onTitleChange={handleTitleChange}
-        >
-          <Tab key="tab1" title="Tab 1">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      const tab = getByRole('tab', { name: 'Tab 1' });
-      const menuTrigger = tab.parentElement?.querySelector(
-        'button[aria-haspopup="true"]',
-      ) as HTMLButtonElement;
-
-      expect(menuTrigger).toBeTruthy();
-      await user.click(menuTrigger);
-
-      const renameItem = await findByRole('menuitem', { name: 'Rename' });
-      await user.click(renameItem);
-
-      // The InlineInput should still be in edit mode after the menu closes
-      // and its FocusScope tries to restore focus to the trigger.
-      const input = await waitFor(() => {
-        const el = queryByRole('textbox');
-        expect(el).toBeInTheDocument();
-
-        return el as HTMLInputElement;
-      });
-
-      // Simulate the focus-theft race directly: explicitly steal focus to
-      // the menu trigger right after the input mounts. With the InlineInput
-      // grace period + TabButton refocus pass, the rename session should
-      // survive and not commit-on-blur.
-      act(() => {
-        menuTrigger.focus();
-      });
-
-      await act(async () => {
-        await new Promise((resolve) => requestAnimationFrame(resolve));
-      });
-
-      expect(queryByRole('textbox')).toBeInTheDocument();
-      expect(handleTitleChange).not.toHaveBeenCalled();
-    });
-
-    // Regression: with `contextMenu="context-only"` and no delete handler
-    // / custom actions, the `TabContainer` ref is replaced by the context
-    // menu's target ref and the `Actions` element is never rendered — so
-    // both `containerRef` and `actionsRef` are null. The TabButton refocus
-    // pass must still be able to locate the editing input (via the tab
-    // button ref, which is always wired up).
-    it('also mounts the rename input in context-only mode with no actions', async () => {
-      const handleTitleChange = vi.fn();
-      const { findByRole, getByRole, queryByRole } = renderWithRoot(
-        <Tabs
-          isEditable
-          contextMenu="context-only"
-          defaultActiveKey="tab1"
-          menu={<Menu.Item key="rename">Rename</Menu.Item>}
-          onTitleChange={handleTitleChange}
-        >
-          <Tab key="tab1" title="Tab 1">
-            Content 1
-          </Tab>
-        </Tabs>,
-      );
-
-      const tab = getByRole('tab', { name: 'Tab 1' });
-
-      // Open the context menu via right-click and select Rename.
-      await act(async () => {
-        fireEvent.contextMenu(tab);
-      });
-      const renameItem = await findByRole('menuitem', { name: 'Rename' });
-      await act(async () => {
-        fireEvent.click(renameItem);
-      });
-
-      await waitFor(() => {
-        expect(queryByRole('textbox')).toBeInTheDocument();
-      });
-
-      // Let the TabButton refocus pass run (rAF). With the fix, this finds
-      // the editing input via the tab button ref even though both
-      // `containerRef` and `actionsRef` are null in this configuration.
-      await act(async () => {
-        await new Promise((resolve) => requestAnimationFrame(resolve));
-      });
-
-      expect(queryByRole('textbox')).toBeInTheDocument();
-      expect(handleTitleChange).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('Context menu modes', () => {
-    const menu = (
-      <>
-        <Menu.Item key="rename">Rename</Menu.Item>
-      </>
-    );
-
-    /** Tab row renders actions outside Item; Item also has an empty Actions slot. */
-    function actionsContainerWithButtons(tab: HTMLElement) {
-      let node: HTMLElement | null = tab.parentElement;
-      while (node) {
-        for (const child of Array.from(node.children)) {
-          if (
-            child.getAttribute('data-element') === 'Actions' &&
-            child.querySelector('button')
-          ) {
-            return child as HTMLElement;
-          }
-        }
-        node = node.parentElement;
-      }
-      return null;
-    }
-
-    it('contextMenu true keeps overflow trigger and hides inline close when menu is non-empty', () => {
-      const onDelete = vi.fn();
-      const { getByRole } = renderWithRoot(
-        <Tabs
-          contextMenu
-          defaultActiveKey="tab1"
-          menu={menu}
-          onDelete={onDelete}
-        >
-          <Tab key="tab1" title="Tab 1">
-            Content
-          </Tab>
-        </Tabs>,
-      );
-
-      const actionsEl = actionsContainerWithButtons(
-        getByRole('tab', { name: 'Tab 1' }),
-      );
-      expect(actionsEl).toBeTruthy();
-      expect(
-        within(actionsEl!).queryByRole('button', {
-          name: 'Close',
-          hidden: true,
-        }),
-      ).not.toBeInTheDocument();
-      expect(within(actionsEl!).getAllByRole('button').length).toBeGreaterThan(
-        0,
-      );
-    });
-
-    it('contextMenu context-only shows inline close and no overflow trigger', () => {
-      const onDelete = vi.fn();
-      const { getByRole } = renderWithRoot(
-        <Tabs
-          defaultActiveKey="tab1"
-          contextMenu="context-only"
-          menu={menu}
-          onDelete={onDelete}
-        >
-          <Tab key="tab1" title="Tab 1">
-            Content
-          </Tab>
-        </Tabs>,
-      );
-
-      const actionsEl = actionsContainerWithButtons(
-        getByRole('tab', { name: 'Tab 1' }),
-      );
-      expect(actionsEl).toBeTruthy();
-      expect(
-        within(actionsEl!).getByRole('button', { name: 'Close' }),
-      ).toBeInTheDocument();
-      expect(within(actionsEl!).queryAllByRole('button')).toHaveLength(1);
-    });
-
-    it('contextMenu context-only with no menu behaves like false (inline close, no context menu)', () => {
-      const onDelete = vi.fn();
-      const { getByRole, container } = renderWithRoot(
-        <Tabs
-          defaultActiveKey="tab1"
-          contextMenu="context-only"
-          onDelete={onDelete}
-        >
-          <Tab key="tab1" title="Tab 1">
-            Content
-          </Tab>
-        </Tabs>,
-      );
-
-      const actionsEl = actionsContainerWithButtons(
-        getByRole('tab', { name: 'Tab 1' }),
-      );
-      expect(actionsEl).toBeTruthy();
-      expect(
-        within(actionsEl!).getByRole('button', { name: 'Close' }),
-      ).toBeInTheDocument();
-      expect(within(actionsEl!).queryAllByRole('button')).toHaveLength(1);
-      expect(
-        container.querySelector('[aria-label="Open context menu"]'),
-      ).not.toBeInTheDocument();
-    });
-  });
-
   describe('hideTabListScroll', () => {
     it('should not render Scrollbar element when hideTabListScroll is true', () => {
       const { container } = renderWithRoot(
